Adjust code structure
