+++ examples/pytorch/llm/src/utils/__init__.py
@@ -1,10 +1,9 @@
-from .argument import InferArguments, SftArguments
+from .argument import InferArguments, SftArguments, select_bnb, select_dtype
 from .dataset import DATASET_MAPPING, get_dataset
-from .metric_utils import compute_nlg_metrics
 from .model import MODEL_MAPPING, get_model_tokenizer
 from .preprocess import TEMPLATE_MAPPING, get_preprocess
 from .utils import (broadcast_string, check_json_format, dataset_map,
                     download_dataset, find_all_linear_for_lora,
                     get_dist_setting, inference, is_ddp_plus_mp, is_dist,
-                    is_local_master, is_master, plot_images, select_bnb,
-                    select_dtype, show_layers, sort_by_max_length)
+                    is_local_master, is_master, plot_images, show_layers,
+                    sort_by_max_length)
