+++ examples/pytorch/llm/src/utils/utils.py
@@ -27,197 +27,14 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from tqdm.auto import tqdm
 from transformers import GenerationConfig, TextStreamer, trainer
 
-from swift import get_logger
 from swift.hub import ModelScopeConfig
-from swift.utils.tb_utils import (TB_COLOR, TB_COLOR_SMOOTH,
-                                  read_tensorboard_file, tensorboard_smoothing)
-from .callback import DefaultFlowCallbackNew, ProgressCallbackNew
+from swift.utils import (get_logger, is_ddp_plus_mp, is_dist, is_local_master,
+                         is_master)
 
 logger = get_logger()
 ms_logger = get_ms_logger()
 
 os.environ['TOKENIZERS_PARALLELISM'] = 'true'
-
-DTYPE_MAPPING = {
-    'fp16': torch.float16,
-    'bf16': torch.bfloat16,
-    'fp32': torch.float32
-}
-
-
-def get_dist_setting() -> Tuple[int, int, int, int]:
-    """return rank, local_rank, world_size"""
-    rank = int(os.getenv('RANK', -1))
-    local_rank = int(os.getenv('LOCAL_RANK', -1))
-    world_size = int(os.getenv('WORLD_SIZE', 1))
-    local_world_size = int(os.getenv('LOCAL_WORLD_SIZE', 1))
-    return rank, local_rank, world_size, local_world_size
-
-
-def is_master():
-    rank = get_dist_setting()[0]
-    return rank in {-1, 0}
-
-
-def is_local_master():
-    local_rank = get_dist_setting()[1]
-    return local_rank in {-1, 0}
-
-
-def is_dist():
-    """Determine if the training is distributed"""
-    rank, local_rank, _, _ = get_dist_setting()
-    return rank >= 0 and local_rank >= 0
-
-
-def show_layers(model: Module, max_lines: Optional[int] = 20) -> None:
-    named_p = list(model.named_parameters())
-    for i, (n, p) in enumerate(named_p):
-        if max_lines is not None and i >= max_lines:
-            logger.info('...')
-            break
-        logger.info(
-            f'[{n}]: requires_grad={p.requires_grad}, dtype={p.dtype}, device={p.device}'
-        )
-
-
-def plot_images(images_dir: str,
-                tb_dir: str,
-                smooth_key: List[str],
-                smooth_val: float = 0.9,
-                figsize: Tuple[int, int] = (8, 5),
-                dpi: int = 100) -> None:
-    """Using tensorboard's data content to plot images"""
-    os.makedirs(images_dir, exist_ok=True)
-    fname = [
-        fname for fname in os.listdir(tb_dir)
-        if os.path.isfile(os.path.join(tb_dir, fname))
-    ][0]
-    tb_path = os.path.join(tb_dir, fname)
-    data = read_tensorboard_file(tb_path)
-
-    for k in data.keys():
-        _data = data[k]
-        steps = [d['step'] for d in _data]
-        values = [d['value'] for d in _data]
-        if len(values) == 0:
-            continue
-        _, ax = plt.subplots(1, 1, squeeze=True, figsize=figsize, dpi=dpi)
-        ax.set_title(k)
-        if len(values) == 1:
-            ax.scatter(steps, values, color=TB_COLOR_SMOOTH)
-        elif k in smooth_key:
-            ax.plot(steps, values, color=TB_COLOR)
-            values_s = tensorboard_smoothing(values, smooth_val)
-            ax.plot(steps, values_s, color=TB_COLOR_SMOOTH)
-        else:
-            ax.plot(steps, values, color=TB_COLOR_SMOOTH)
-        fpath = os.path.join(images_dir, k.replace('/', '_'))
-        plt.savefig(fpath, dpi=dpi, bbox_inches='tight')
-
-
-def inference(input_ids: List[int],
-              model,
-              tokenizer,
-              streamer: Optional[TextStreamer] = None,
-              skip_prompt: bool = False) -> str:
-    generation_config = getattr(model, 'generation_config', None)
-    if not skip_prompt:
-        print(f'[INFERENCE]{tokenizer.decode(input_ids)}', end='')
-    input_ids = torch.tensor(input_ids)[None].cuda()
-    attention_mask = torch.ones_like(input_ids)
-    model.eval()
-    generate_ids = model.generate(
-        input_ids=input_ids,
-        attention_mask=attention_mask,
-        streamer=streamer,
-        generation_config=generation_config)
-    output_text = tokenizer.decode(generate_ids[0, len(input_ids[0]):])
-    if streamer is None:
-        print(output_text)
-    return output_text
-
-
-def select_dtype(dtype: str) -> Tuple[Dtype, bool, bool]:
-    """
-    dtype: Literal['fp16', 'bf16', 'fp32']
-    """
-    torch_dtype = DTYPE_MAPPING[dtype]
-
-    assert torch_dtype in {torch.float16, torch.bfloat16, torch.float32}
-    if torch_dtype == torch.float16:
-        fp16, bf16 = True, False
-    elif torch_dtype == torch.bfloat16:
-        support_bf16 = torch.cuda.is_bf16_supported()
-        if not support_bf16:
-            logger.warning(f'support_bf16: {support_bf16}')
-        fp16, bf16 = False, True
-    else:
-        fp16, bf16 = False, False
-    return torch_dtype, fp16, bf16
-
-
-def select_bnb(quantization_bit: int,
-               bnb_4bit_compute_dtype: str) -> Tuple[Dtype, bool, bool]:
-    bnb_4bit_compute_dtype = DTYPE_MAPPING[bnb_4bit_compute_dtype]
-    assert bnb_4bit_compute_dtype in {
-        torch.float16, torch.bfloat16, torch.float32
-    }
-    if quantization_bit == 4:
-        load_in_4bit, load_in_8bit = True, False
-    elif quantization_bit == 8:
-        load_in_4bit, load_in_8bit = False, True
-    else:
-        load_in_4bit, load_in_8bit = False, False
-
-    return bnb_4bit_compute_dtype, load_in_4bit, load_in_8bit
-
-
-def broadcast_string(string: Optional[str], buffer_size: int = 200) -> str:
-    """String broadcasting in case of DDP
-    string: main rank: str
-        other rank: None
-    return: all rank: str
-    """
-    assert dist.is_initialized()
-    rank, local_rank, _, _ = get_dist_setting()
-    assert rank >= 0
-    if rank == 0:
-        assert string is not None
-        tensor = torch.tensor(
-            [ord(c) for c in string] + [0] * (buffer_size - len(string)),
-            dtype=torch.int64,
-            device=local_rank)
-    else:
-        tensor = torch.zeros(buffer_size, dtype=torch.int64, device=local_rank)
-    dist.broadcast(tensor, 0)
-    first_zero = (tensor == 0).nonzero()[0].item()
-    res = tensor.tolist()[:first_zero]
-    return ''.join([chr(x) for x in res])
-
-
-def find_all_linear_for_lora(model: Module,
-                             quantization_bit: int,
-                             model_type: Optional[str] = None) -> List[str]:
-    """ref: https://github.com/artidoro/qlora"""
-    head_module_name = 'lm_head'
-    if model_type.startswith('chatglm2-6b'):
-        head_module_name = 'output_layer'
-    if quantization_bit == 4:
-        from bitsandbytes.nn import Linear4bit
-        linear_cls = Linear4bit
-    elif quantization_bit == 8:
-        from bitsandbytes.nn import Linear8bitLt
-        linear_cls = Linear8bitLt
-    else:
-        linear_cls = Linear
-    lora_module_names = set()
-    for name, module in model.named_modules():
-        if isinstance(module, linear_cls):
-            module_name = name.split('.')[-1]
-            if head_module_name not in module_name:
-                lora_module_names.add(module_name)
-    return list(lora_module_names)
 
 
 def download_dataset(model_id: str,
@@ -250,36 +67,6 @@ def download_files(url: str, local_path: str, cookies) -> None:
             f.write(data)
 
 
-def sort_by_max_length(dataset: HfDataset, num_dataset: int) -> HfDataset:
-    dataset_len = [len(d['input_ids']) for d in tqdm(dataset)]
-    idx = heapq.nlargest(
-        num_dataset, range(len(dataset_len)), key=lambda i: dataset_len[i])
-    input_ids = []
-    labels = []
-    for i in tqdm(idx):
-        input_ids.append(dataset[i]['input_ids'])
-        labels.append(dataset[i]['labels'])
-    return HfDataset.from_dict({'input_ids': input_ids, 'labels': labels})
-
-
-def check_json_format(obj: Any) -> Any:
-    if obj is None or isinstance(
-            obj, (int, float, str, complex)):  # bool is a subclass of int
-        return obj
-
-    if isinstance(obj, Sequence):
-        res = []
-        for x in obj:
-            res.append(check_json_format(x))
-    elif isinstance(obj, Mapping):
-        res = {}
-        for k, v in obj.items():
-            res[k] = check_json_format(v)
-    else:
-        res = repr(obj)  # e.g. function
-    return res
-
-
 _old_msdataset_load = MsDataset.load
 
 
@@ -294,18 +81,6 @@ def _msdataset_ddp_load(*args, **kwargs):
     if is_dist():
         dist.barrier()
     return dataset
-
-
-def is_ddp_plus_mp() -> bool:
-    if not is_dist():
-        return False
-    n_gpu = torch.cuda.device_count()
-    local_world_size = get_dist_setting()[3]
-    assert n_gpu % local_world_size == 0
-    if n_gpu // local_world_size >= 2:
-        logger.info('Using DDP + MP(device_map)')
-        return True
-    return False
 
 
 def _get_max_memory(device_ids: List[int]) -> Dict[Union[int, str], int]:
@@ -393,8 +168,6 @@ else:
     ms_logger.setLevel(logging.ERROR)
 
 # monkey patching
-trainer.DEFAULT_PROGRESS_CALLBACK = ProgressCallbackNew
-trainer.DEFAULT_CALLBACKS = [DefaultFlowCallbackNew]
 MsDataset.load = _msdataset_ddp_load
 if is_ddp_plus_mp():
     import transformers
