+++ swift/utils/__init__.py
@@ -1,7 +1,15 @@
 # Copyright (c) Alibaba, Inc. and its affiliates.
 
+from .llm_utils import (data_collate_fn, find_all_linear_for_lora, inference,
+                        lower_bound, print_example, sort_by_max_length,
+                        stat_dataset)
 from .logger import get_logger
-from .torch_utils import (add_version_to_work_dir, broadcast_string,
-                          get_dist_setting, get_seed, is_dist, is_local_master,
-                          is_master, is_on_same_device, parse_args,
-                          print_model_info, seed_everything, show_layers)
+from .metric import compute_nlg_metrics
+from .tb_utils import (TB_COLOR, TB_COLOR_SMOOTH, read_tensorboard_file,
+                       tensorboard_smoothing)
+from .torch_utils import (broadcast_string, get_dist_setting, is_ddp_plus_mp,
+                          is_dist, is_local_master, is_master,
+                          is_on_same_device, print_model_info, seed_everything,
+                          show_layers)
+from .utils import (add_version_to_work_dir, check_json_format, get_seed,
+                    parse_args)
