+++ swift/utils/llm_utils.py
@@ -5,7 +5,9 @@ from typing import Any, Callable, Dict, List, Optional
 import numpy as np
 import torch
 from datasets import Dataset as HfDataset
+from torch.nn import Module
 from torch.nn.utils.rnn import pad_sequence
+from transformers import TextStreamer
 
 from .logger import get_logger
 
@@ -68,3 +70,61 @@ def print_example(example: Dict[str, Any], tokenizer) -> None:
     logger.info(f'[LABLES_IDS] {labels}')
     logger.info(
         f'[LABLES] [-100 * {n_mask}]{tokenizer.decode(labels[n_mask:])}')
+
+
+def find_all_linear_for_lora(model: Module,
+                             quantization_bit: int,
+                             model_type: Optional[str] = None) -> List[str]:
+    """ref: https://github.com/artidoro/qlora"""
+    head_module_name = 'lm_head'
+    if model_type.startswith('chatglm2-6b'):
+        head_module_name = 'output_layer'
+    if quantization_bit == 4:
+        from bitsandbytes.nn import Linear4bit
+        linear_cls = Linear4bit
+    elif quantization_bit == 8:
+        from bitsandbytes.nn import Linear8bitLt
+        linear_cls = Linear8bitLt
+    else:
+        linear_cls = Linear
+    lora_module_names = set()
+    for name, module in model.named_modules():
+        if isinstance(module, linear_cls):
+            module_name = name.split('.')[-1]
+            if head_module_name not in module_name:
+                lora_module_names.add(module_name)
+    return list(lora_module_names)
+
+
+def sort_by_max_length(dataset: HfDataset, num_dataset: int) -> HfDataset:
+    dataset_len = [len(d['input_ids']) for d in tqdm(dataset)]
+    idx = heapq.nlargest(
+        num_dataset, range(len(dataset_len)), key=lambda i: dataset_len[i])
+    input_ids = []
+    labels = []
+    for i in tqdm(idx):
+        input_ids.append(dataset[i]['input_ids'])
+        labels.append(dataset[i]['labels'])
+    return HfDataset.from_dict({'input_ids': input_ids, 'labels': labels})
+
+
+def inference(input_ids: List[int],
+              model,
+              tokenizer,
+              streamer: Optional[TextStreamer] = None,
+              skip_prompt: bool = False) -> str:
+    generation_config = getattr(model, 'generation_config', None)
+    if not skip_prompt:
+        print(f'[INFERENCE]{tokenizer.decode(input_ids)}', end='')
+    input_ids = torch.tensor(input_ids)[None].cuda()
+    attention_mask = torch.ones_like(input_ids)
+    model.eval()
+    generate_ids = model.generate(
+        input_ids=input_ids,
+        attention_mask=attention_mask,
+        streamer=streamer,
+        generation_config=generation_config)
+    output_text = tokenizer.decode(generate_ids[0, len(input_ids[0]):])
+    if streamer is None:
+        print(output_text)
+    return output_text
