+++ swift/utils/metric.py
@@ -1,16 +1,16 @@
 # Copyright (c) Alibaba, Inc. and its affiliates.
 
+import jieba
 import numpy as np
+from nltk.translate.bleu_score import SmoothingFunction, sentence_bleu
+from rouge.rouge import Rouge
 
-from swift import get_logger
+from .logger import get_logger
 
 logger = get_logger()
 
 
 def compute_nlg_metrics(prediction, tokenizer):
-    import jieba
-    from nltk.translate.bleu_score import SmoothingFunction, sentence_bleu
-    from rouge.rouge import Rouge
     preds, labels = prediction[0], prediction[1]
 
     score_dict = {'rouge-1': [], 'rouge-2': [], 'rouge-l': [], 'bleu-4': []}
