+++ swift/utils/torch_utils.py
@@ -12,7 +12,6 @@ import torch
 import torch.distributed as dist
 from numpy.random import RandomState
 from torch.nn import Module
-from transformers import HfArgumentParser
 
 from .logger import get_logger, is_master
 
@@ -33,32 +32,6 @@ def _find_free_port() -> str:
     sock.close()
     # NOTE: there is still a chance the port could be taken by other processes.
     return port
-
-
-def _get_version(work_dir: str) -> int:
-    if os.path.isdir(work_dir):
-        fnames = os.listdir(work_dir)
-    else:
-        fnames = []
-    v_list = [-1]
-    for fname in fnames:
-        m = re.match(r'v(\d+)', fname)
-        if m is None:
-            continue
-        v = m.group(1)
-        v_list.append(int(v))
-    return max(v_list) + 1
-
-
-def add_version_to_work_dir(work_dir: str) -> str:
-    """add version"""
-    work_dir = os.path.abspath(work_dir)
-    version = _get_version(work_dir)
-    time = dt.datetime.now().strftime('%Y%m%d-%H%M%S')
-
-    work_dir = os.path.join(work_dir, f'v{version}-{time}')
-    logger.info(f'work_dir: {work_dir}')
-    return work_dir
 
 
 def seed_everything(seed: Optional[int] = None, gpu_dtm: bool = False) -> int:
@@ -109,23 +82,6 @@ def find_sub_module(module: torch.nn.Module,
     return _modules
 
 
-def get_seed(random_state: RandomState) -> int:
-    seed_max = np.iinfo(np.int32).max
-    seed = random_state.randint(0, seed_max)
-    return seed
-
-
-_T = TypeVar('_T')
-
-
-def parse_args(class_type: Type[_T],
-               argv: Optional[List[str]] = None) -> Tuple[_T, List[str]]:
-    parser = HfArgumentParser([class_type])
-    args, remaining_args = parser.parse_args_into_dataclasses(
-        argv, return_remaining_strings=True)
-    return args, remaining_args
-
-
 def get_dist_setting() -> Tuple[int, int, int, int]:
     """return rank, local_rank, world_size"""
     rank = int(os.getenv('RANK', -1))
@@ -146,6 +102,18 @@ def is_dist():
     return rank >= 0 and local_rank >= 0
 
 
+def is_ddp_plus_mp() -> bool:
+    if not is_dist():
+        return False
+    n_gpu = torch.cuda.device_count()
+    local_world_size = get_dist_setting()[3]
+    assert n_gpu % local_world_size == 0
+    if n_gpu // local_world_size >= 2:
+        logger.info('Using DDP + MP(device_map)')
+        return True
+    return False
+
+
 def show_layers(model: Module, max_lines: Optional[int] = 20) -> None:
     named_p = list(model.named_parameters())
     for i, (n, p) in enumerate(named_p):
@@ -157,7 +125,7 @@ def show_layers(model: Module, max_lines: Optional[int] = 20) -> None:
         )
 
 
-def broadcast_string(string: Optional[str], buffer_size: int = 100) -> str:
+def broadcast_string(string: Optional[str], buffer_size: int = 200) -> str:
     """String broadcasting in case of DDP
     string: main rank: str
         other rank: None
