+++ swift/utils/utils.py
@@ -1,0 +1,67 @@
+import os
+from typing import Any, List, Optional, Tuple, Type, TypeVar
+
+import numpy as np
+from numpy.random import RandomState
+from transformers import HfArgumentParser
+
+
+def check_json_format(obj: Any) -> Any:
+    if obj is None or isinstance(
+            obj, (int, float, str, complex)):  # bool is a subclass of int
+        return obj
+
+    if isinstance(obj, Sequence):
+        res = []
+        for x in obj:
+            res.append(check_json_format(x))
+    elif isinstance(obj, Mapping):
+        res = {}
+        for k, v in obj.items():
+            res[k] = check_json_format(v)
+    else:
+        res = repr(obj)  # e.g. function
+    return res
+
+
+def _get_version(work_dir: str) -> int:
+    if os.path.isdir(work_dir):
+        fnames = os.listdir(work_dir)
+    else:
+        fnames = []
+    v_list = [-1]
+    for fname in fnames:
+        m = re.match(r'v(\d+)', fname)
+        if m is None:
+            continue
+        v = m.group(1)
+        v_list.append(int(v))
+    return max(v_list) + 1
+
+
+def add_version_to_work_dir(work_dir: str) -> str:
+    """add version"""
+    work_dir = os.path.abspath(work_dir)
+    version = _get_version(work_dir)
+    time = dt.datetime.now().strftime('%Y%m%d-%H%M%S')
+
+    work_dir = os.path.join(work_dir, f'v{version}-{time}')
+    logger.info(f'work_dir: {work_dir}')
+    return work_dir
+
+
+def get_seed(random_state: RandomState) -> int:
+    seed_max = np.iinfo(np.int32).max
+    seed = random_state.randint(0, seed_max)
+    return seed
+
+
+_T = TypeVar('_T')
+
+
+def parse_args(class_type: Type[_T],
+               argv: Optional[List[str]] = None) -> Tuple[_T, List[str]]:
+    parser = HfArgumentParser([class_type])
+    args, remaining_args = parser.parse_args_into_dataclasses(
+        argv, return_remaining_strings=True)
+    return args, remaining_args
